bfd: Fix initialized data warnings for rs6000 on 64-bit AIX >= 4.3
I got a load of compile warnings about initialized memory in
rs6000-core.c when compiling on an AIX 7.2 POWER machine.

The cause is the CORE_NEW macro. This macro should (according to
the comment) return 1 if a given CoreHdr structure contains data
in the new AIX 4.3+ format. On systems where AIX_CORE_DUMPX_CORE
and BFD64 are true the CoreHdr structure can only contain a new
style core dump so the macro always evaluates to 1. The compiler
doesn't know this and gives a warning. Still, since we know the
macro will always evaluate to 1 we can define it to 1 and fix the
warnings.